Add intra-line scrolling and abstract wrapping away from the window code
Previously, the scroll position was represented by a Y position in
text-space, so we couldn't scroll within a wrapped line. This worked
fine-ish, but with the new prompt implementation became a serious
nuisance, and was kind of silly anyway. This fixes it, by introducing
a window-space (covering the whole file), relative to which cursor
and scroll positions are specified, where coordinates refer to wrapped
lines, instead of the original ones.