Fix bug where the cursor would be misplaced when launching mflg with a specific location in a file
The bug occurred because main resized the app right before initiating
the main event loop, which resized the main window, which clamped the
cursor position to be within the window bounds. This was a remnant of
when we didn't support an offscreen cursor. The clamping and the
initial resize (which was made redundant by the changes to the navigation
code) are now removed.