Reload file correctly when its parent directory is created
This allows mflg to, for instance, detect when another instance of
mflg does this.

The implementation is somewhat hacky; it uses a short timer on
parent directory creation in order to resolve the race between
the file being created and us reloading and re-watching it, after
we get the notification. This seems to be the best we can do
at the moment.