Use our own file-watching code
The new internal/pathwatch package offers an easier to use API for our use case.
For now, we use a polling implementation, which is simple and portable but inefficient.

The motivation for this is to make it easier to watch multiple files simulatenously -
specifically, watch both the file we're editing as well as the configuration file, to
enable configuration auto-reloading.